Friendbot, comments: move actions back into services/horizon/internal
