+++ services/friendbot/internal/actions_friendbot.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	client "github.com/stellar/go/clients/horizon"
-	"github.com/stellar/go/services/horizon/actions"
+	"github.com/stellar/go/services/horizon/internal/actions"
 	"github.com/stellar/go/services/horizon/render/hal"
 	"github.com/stellar/go/services/horizon/render/problem"
 	"github.com/zenazn/goji/web"
